Initial version of the depends-on plugin
The depends-on plugin for Gerrit is designed for groups using
the 'Depends-on:' change comment to manage inter-project change
dependencies.

When a new change is created via the cherry-pick API, this plugin
propagates dependencies by copying the 'Depends-on:' comment from
the source change to the new destination change.

NOTE: Only the latest 'Depends-on:' comment is considered and the
      older ones are ignored.

This plugin builds against Gerrit 3.2.10(using bazel) and includes
docker based functional tests.

Change-Id: I9a4b8b1499422464310cd6fd54e01fe0d1cf6714