Fix running docker tests in-tree
Instead of using the hardcoded Gerrit docker image from run.sh, we want
to use the gerrit.war from core that will exactly match the API we build
against.

To make this work and retain support for using the 'docker-tests' target
in a standalone build, we add a defined config setting for the build
mode and use the standalone .bazelrc to set it. Then we assume that an
unset config means we're doing an in-tree build.

It would be nice to have this config set by bazlets, but that's going to
be more complicated.

Change-Id: I533fa8faff49c1439348780199fdad43bad47b07