add polymer UI to view and edit dependencies
Add a polymer UI that uses the --depends-on--all query results to
display dependencies. Users are also allowed to edit the dependencies
by clicking on the edit icon. The edit window posts a new change level
depends-on comment to update the depends-on. Update documentation and
comment validation error message to point to the new UI.

Add eslint so that javascript files can be linted using google
javascript formatting guidelines. eslint can be ran with following
commands:
  $ bazel test lint_test --test_output=errors
  $ bazel run lint_test -- --fix /path/to/file.js

Change-Id: I7a1870e48a83d8e4927274aaacd42c4b9dc5ea86