Retry review on lock failures
The review api is prone to lock failures as there will be a contention
to lock the notedb ref. Retrying on lock failures will help mitigate
this. Retries have already been implemented in REST API code path at
a high enough layer (RestApiServlet) so that it applies for all REST
APIs. But when invoking review api internally from a plugin won't
enable the retires defined in RestApiServlet.

Change-Id: Ifa9d9a80b1be56c755b7d248941a9404f84dc1a5