test/docker/run: Stop caching artifacts
Trying to cache artifacts like this causes surprises where a different
version of plugin or gerrit is being tested than expected.

We don't expect a --gerrit-war often during normal plugin development,
but add a default of bazel-bin/depends-on.jar for the plugin jar option
so that it doesn't have to be specified to run tests.

Using the 'bazel test docker-tests' command is still better because it
ensures the plugin version you're testing matches the tree you have
checked out.

Change-Id: I9328d12851e7dfc04a5ec4b7b18e87fcb1e20753