Add in_depends-on operator
Add in_depends-on operator that returns "Depends-on" change
dependencies specified in the comments of the provided change.

To use "in_depends-on" operator change operator aliasing is
needed since query parser is not able to parse dash(-) in depends-on
operator.

Sample gerrit configuration for in_depends-on operator aliasing:

[operator-alias "change"]
  independson = in_depends-on

Sample query command using 'in_depends-on' operator alias
  ssh -p 29418 localhost gerrit query independson:<change>

Change-Id: I50c53a54bd6b6bb9c49dea3de3b6e30a7767ad98