Add support for --depends-on--all query option
Add --depends-on--all option to gerrit query SSH command and REST API.
This option will add a plugin section to the output which displays all
the depends-ons from the latest Depends-on comment on the change. This
option helps in building a depends-on web UI in future. Also add
functional tests and documentation.

Change-Id: I713f7765f47d115d499299f449fe08a70dd1a532