Make InDependsOnOperator like a ChangeDataSource
To make InDependsOnOperator behave like a ChangesDataSource, make it
return an ORed list of ChangeIndexPredicates which will be processed as
a ChangeDataSource. Although ORing together a bunch of
ChangeIndexPredicates results in slower match()s then the previous
implementation, this is negligible with only a few changes (likely
unmeasurable in real life), as a ChangeDataSource it will not have to
get ANDed with status predicates when unqualifed, and even better, it
will get selected when ANDed with status sources as the source.

Change-Id: I0d4487ceef47e8b8416ea8403072d331de4d9d88