Cache depends on by change per thread
Caching depends on by change per thread improves the performance
of queries such as "change:<> has:a_depends-on --depends-on--all" by
3ms against ES index backend (i.e. on avg took 20ms before and 17ms
after this change over 50 samples). This improvement is due to the fact
that has:a_depends-on and --depends-on--all computes the same
depends-on information.

One can also benefit from this performance improvement when using
has:a_depends-on and in_depends-on:<> operators in Task plugin config.

Change-Id: Ia97f08bf8c0e607a4044db281f191bdd7ebe3f03