test: Fix docker-tests for in-tree/master
When running the docker tests against master we need to depend on the
in-tree WAR so that the runtime APIs match what we compiled against. In
order to depend on the WAR we need to add the polymer-bridges project
(core submodule) as a required project so that the core build is
successful.

This also requires us to remove all other plugins from the 3.9 docker
image to avoid any incompatibilities they may have with master.

As a result of this change, the bazel test docker-tests target now only
works with an in-tree build on the master branch. This is an unfortunate
difference compared to the stable branches, but it makes sense given
that we need a Gerrit WAR with compatible runtime APIs running in the
container.

Change-Id: I533fa8faff49c1439348780199fdad43bad47b07