UI: Move depends-on field above submit-requirements
The intent is to move it closer to other core metadata items, so
that the change-info section does not look disjointed. This change
also adjusts the edit dialog to place it over other elements and
to use fonts from core for the text area.

Also, change the title name of then field from 'Depends-on' to 'DependsOn' as the original name is too long and will wrap into
the next line when the core fields have a shorter title width.

Screenshots: https://imgur.com/a/1wDNSEp

Change-Id: If44fd7ac39ec498bdc115701f877d3ea4e85be62