UI: Move depends-on field above submit-requirements
The intent is to move it closer to other core metadata items, so
that the change-info section does not look disjointed. This change
also adjusts the edit dialog to place it over other elements and
to use fonts from core for the text area.

Screenshots: https://imgur.com/a/1wDNSEp

Change-Id: If44fd7ac39ec498bdc115701f877d3ea4e85be62