Create an 'extensions' subpackage
This will allow other plugins providing dependency resolution
implementations a way to include the build dependency through Bazel.

Since standalone and in-tree builds will have a different absolute
package path for the new subpackage, use a bit of code from the
bazel-skylib repo to do the equivalent of os.path.join() with the
package name.

Change-Id: I199a8488e612177a2ad03f95a95f35e0c2fbb37c