UI: Render direct url to dependent change
When the dependent change is an int, render a direct url to it to avoid
an extra click. Currently, when users click a dependent change in the
UI, they are directed to a change-list view, requiring an additional
click to view the change. If the dependency is not an int (for example,
a change-id), we continue to render the query url.

Change-Id: I98f858dfc7a524ba81600bf59c9fbb116b1db440